[CompilerPass] move basic Compiler Pass for Registry into Registry component, not every bundle has a dependency to PimcoreBundle for that
