+++ src/CoreShop/Bundle/CoreBundle/DependencyInjection/Compiler/RegisterReportsPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\CoreBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\RegisterSimpleRegistryTypePass;
+use CoreShop\Component\Registry\RegisterSimpleRegistryTypePass;
 
 class RegisterReportsPass extends RegisterSimpleRegistryTypePass
 {
