+++ src/CoreShop/Bundle/CurrencyBundle/DependencyInjection/Compiler/CompositeCurrencyContextPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\CurrencyBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\PrioritizedCompositeServicePass;
+use CoreShop\Component\Registry\PrioritizedCompositeServicePass;
 
 final class CompositeCurrencyContextPass extends PrioritizedCompositeServicePass
 {
