+++ src/CoreShop/Bundle/CustomerBundle/DependencyInjection/Compiler/CompositeCustomerContextPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\CustomerBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\PrioritizedCompositeServicePass;
+use CoreShop\Component\Registry\PrioritizedCompositeServicePass;
 
 final class CompositeCustomerContextPass extends PrioritizedCompositeServicePass
 {
