+++ src/CoreShop/Bundle/IndexBundle/DependencyInjection/Compiler/RegisterGetterPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\IndexBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\RegisterRegistryTypePass;
+use CoreShop\Component\Registry\RegisterRegistryTypePass;
 
 class RegisterGetterPass extends RegisterRegistryTypePass
 {
