+++ src/CoreShop/Bundle/OrderBundle/DependencyInjection/Compiler/PurchasableDiscountCalculatorsPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\OrderBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\RegisterSimpleRegistryTypePass;
+use CoreShop\Component\Registry\RegisterSimpleRegistryTypePass;
 
 final class PurchasableDiscountCalculatorsPass extends RegisterSimpleRegistryTypePass
 {
