+++ src/CoreShop/Bundle/OrderBundle/DependencyInjection/Compiler/PurchasableRetailPriceCalculatorsPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\OrderBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\RegisterSimpleRegistryTypePass;
+use CoreShop\Component\Registry\RegisterSimpleRegistryTypePass;
 
 final class PurchasableRetailPriceCalculatorsPass extends RegisterSimpleRegistryTypePass
 {
