+++ src/CoreShop/Bundle/OrderBundle/DependencyInjection/Compiler/RegisterCartContextsPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\OrderBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\PrioritizedCompositeServicePass;
+use CoreShop\Component\Registry\PrioritizedCompositeServicePass;
 
 final class RegisterCartContextsPass extends PrioritizedCompositeServicePass
 {
