+++ src/CoreShop/Bundle/OrderBundle/DependencyInjection/Compiler/RegisterCartProcessorPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\OrderBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\PrioritizedCompositeServicePass;
+use CoreShop\Component\Registry\PrioritizedCompositeServicePass;
 
 final class RegisterCartProcessorPass extends PrioritizedCompositeServicePass
 {
