+++ src/CoreShop/Bundle/PimcoreBundle/DependencyInjection/Compiler/RegisterSimpleRegistryTypePass.php
@@ -12,79 +12,15 @@
 
 namespace CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler;
 
-use CoreShop\Component\Registry\PrioritizedServiceRegistry;
-use CoreShop\Component\Registry\PrioritizedServiceRegistryInterface;
-use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
-use Symfony\Component\DependencyInjection\Container;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Reference;
+use CoreShop\Component\Registry\RegisterSimpleRegistryTypePass as NewRegisterSimpleRegistryTypePass;
 
-abstract class RegisterSimpleRegistryTypePass implements CompilerPassInterface
-{
+if (class_exists(NewRegisterSimpleRegistryTypePass::class)) {
+    @trigger_error('Class CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\RegisterSimpleRegistryTypePass is deprecated since version 2.2.6 and will be removed in 3.0.0. Use CoreShop\Component\Registry\RegisterRegistryTypePass class instead.', E_USER_DEPRECATED);
+} else {
     /**
-     * @var string
+     * @deprecated Class CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\RegisterSimpleRegistryTypePass is deprecated since version 2.2.6 and will be removed in 3.0.0. Use CoreShop\Component\Registry\RegisterRegistryTypePass class instead.
      */
-    protected $registry;
-
-    /**
-     * @var string
-     */
-    protected $parameter;
-
-    /**
-     * @var string
-     */
-    protected $tag;
-
-    /**
-     * @param string $registry
-     * @param string $parameter
-     * @param string $tag
-     */
-    public function __construct($registry, $parameter, $tag)
+    class RegisterSimpleRegistryTypePass extends NewRegisterSimpleRegistryTypePass
     {
-        $this->registry = $registry;
-        $this->parameter = $parameter;
-        $this->tag = $tag;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function process(ContainerBuilder $container)
-    {
-        if (!$container->has($this->registry)) {
-            return;
-        }
-
-        $registry = $container->getDefinition($this->registry);
-        $registryInterfaces = class_implements($registry->getClass());
-        $isPrioritizedRegistry = false;
-
-        if ($registryInterfaces && in_array(PrioritizedServiceRegistryInterface::class, $registryInterfaces, true)) {
-            $isPrioritizedRegistry = true;
-        }
-
-        $map = [];
-        foreach ($container->findTaggedServiceIds($this->tag) as $id => $attributes) {
-            $definition = $container->findDefinition($id);
-
-            foreach ($attributes as $tag) {
-                if (!isset($tag['type'])) {
-                    $tag['type'] = Container::underscore(substr(strrchr($definition->getClass(), '\\'), 1));
-                }
-
-                $map[$tag['type']] = $tag['type'];
-
-                if ($isPrioritizedRegistry) {
-                    $registry->addMethodCall('register', [$tag['type'], $tag['priority'] ?? 1000, new Reference($id)]);
-                }
-                else {
-                    $registry->addMethodCall('register', [$tag['type'], new Reference($id)]);
-                }
-            }
-        }
-
-        $container->setParameter($this->parameter, $map);
     }
 }
