+++ src/CoreShop/Bundle/ProductBundle/DependencyInjection/Compiler/ProductPriceRuleConditionPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\ProductBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\RegisterRegistryTypePass;
+use CoreShop\Component\Registry\RegisterRegistryTypePass;
 
 final class ProductPriceRuleConditionPass extends RegisterRegistryTypePass
 {
