+++ src/CoreShop/Bundle/ProductBundle/DependencyInjection/Compiler/ProductValidPriceRuleFetcherPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\ProductBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\RegisterSimpleRegistryTypePass;
+use CoreShop\Component\Registry\RegisterSimpleRegistryTypePass;
 
 final class ProductValidPriceRuleFetcherPass extends RegisterSimpleRegistryTypePass
 {
