+++ src/CoreShop/Bundle/ProductQuantityPriceRulesBundle/DependencyInjection/Compiler/ProductQuantityPriceRulesCalculatorPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\ProductQuantityPriceRulesBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\RegisterSimpleRegistryTypePass;
+use CoreShop\Component\Registry\RegisterSimpleRegistryTypePass;
 
 final class ProductQuantityPriceRulesCalculatorPass extends RegisterSimpleRegistryTypePass
 {
