+++ src/CoreShop/Bundle/ProductQuantityPriceRulesBundle/DependencyInjection/Compiler/ProductQuantityPriceRulesConditionPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\ProductQuantityPriceRulesBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\RegisterRegistryTypePass;
+use CoreShop\Component\Registry\RegisterRegistryTypePass;
 
 final class ProductQuantityPriceRulesConditionPass extends RegisterRegistryTypePass
 {
