+++ src/CoreShop/Bundle/ResourceBundle/DependencyInjection/Compiler/RegisterInstallersPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\ResourceBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\RegisterSimpleRegistryTypePass;
+use CoreShop\Component\Registry\RegisterSimpleRegistryTypePass;
 
 final class RegisterInstallersPass extends RegisterSimpleRegistryTypePass
 {
