+++ src/CoreShop/Bundle/RuleBundle/DependencyInjection/Compiler/RuleAvailabilityAssessorPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\RuleBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\RegisterSimpleRegistryTypePass;
+use CoreShop\Component\Registry\RegisterSimpleRegistryTypePass;
 
 class RuleAvailabilityAssessorPass extends RegisterSimpleRegistryTypePass
 {
