+++ src/CoreShop/Bundle/SEOBundle/DependencyInjection/Compiler/ExtractorRegistryServicePass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\SEOBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\RegisterSimpleRegistryTypePass;
+use CoreShop\Component\Registry\RegisterSimpleRegistryTypePass;
 
 final class ExtractorRegistryServicePass extends RegisterSimpleRegistryTypePass
 {
