+++ src/CoreShop/Bundle/StoreBundle/DependencyInjection/Compiler/CompositeStoreContextPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\StoreBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\PrioritizedCompositeServicePass;
+use CoreShop\Component\Registry\PrioritizedCompositeServicePass;
 
 final class CompositeStoreContextPass extends PrioritizedCompositeServicePass
 {
