+++ src/CoreShop/Bundle/ThemeBundle/DependencyInjection/Compiler/CompositeThemeResolverPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\ThemeBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\PrioritizedCompositeServicePass;
+use CoreShop\Component\Registry\PrioritizedCompositeServicePass;
 
 final class CompositeThemeResolverPass extends PrioritizedCompositeServicePass
 {
