+++ src/CoreShop/Bundle/TrackingBundle/DependencyInjection/Compiler/TrackingExtractorPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\TrackingBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\RegisterSimpleRegistryTypePass;
+use CoreShop\Component\Registry\RegisterSimpleRegistryTypePass;
 
 final class TrackingExtractorPass extends RegisterSimpleRegistryTypePass
 {
