+++ src/CoreShop/Bundle/TrackingBundle/composer.json
@@ -19,7 +19,7 @@
   ],
   "require": {
     "php": "^7.2",
-    "coreshop/registry": "^2.0",
+    "coreshop/registry": "^2.2.6",
     "coreshop/tracking": "^2.0",
     "pimcore/pimcore": "^6.6.0"
   },
