+++ src/CoreShop/Component/Registry/RegisterRegistryTypePass.php
@@ -1,0 +1,92 @@
+<?php
+/**
+ * CoreShop.
+ *
+ * This source file is subject to the GNU General Public License version 3 (GPLv3)
+ * For the full copyright and license information, please view the LICENSE.md and gpl-3.0.txt
+ * files that are distributed with this source code.
+ *
+ * @copyright  Copyright (c) 2015-2020 Dominik Pfaffenbauer (https://www.pfaffenbauer.at)
+ * @license    https://www.coreshop.org/license     GNU General Public License version 3 (GPLv3)
+ */
+
+namespace CoreShop\Component\Registry;
+
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\Container;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Reference;
+
+abstract class RegisterRegistryTypePass implements CompilerPassInterface
+{
+    /**
+     * @var string
+     */
+    protected $registry;
+
+    /**
+     * @var string
+     */
+    protected $formRegistry;
+
+    /**
+     * @var string
+     */
+    protected $parameter;
+
+    /**
+     * @var string
+     */
+    protected $tag;
+
+    /**
+     * @param string $registry
+     * @param string $formRegistry
+     * @param string $parameter
+     * @param string $tag
+     */
+    public function __construct($registry, $formRegistry, $parameter, $tag)
+    {
+        $this->registry = $registry;
+        $this->formRegistry = $formRegistry;
+        $this->parameter = $parameter;
+        $this->tag = $tag;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function process(ContainerBuilder $container)
+    {
+        if (!$container->has($this->registry) || !$container->has($this->formRegistry)) {
+            return;
+        }
+
+        $registry = $container->getDefinition($this->registry);
+        $formRegistry = $container->getDefinition($this->formRegistry);
+
+        $map = [];
+        foreach ($container->findTaggedServiceIds($this->tag) as $id => $attributes) {
+            foreach ($attributes as $tag) {
+                $definition = $container->findDefinition($id);
+
+                if (!isset($tag['type'])) {
+                    $tag['type'] = Container::underscore(substr(strrchr($definition->getClass(), '\\'), 1));
+                }
+
+                $map[$tag['type']] = $tag['type'];
+
+                $registry->addMethodCall('register', [$tag['type'], new Reference($id)]);
+
+                if (isset($tag['form-type'])) {
+                    $formRegistry
+                        ->addMethodCall('add', [$tag['type'], 'default', $tag['form-type']]);
+                }
+            }
+        }
+
+        $container->setParameter($this->parameter, $map);
+    }
+}
+
+class_alias(RegisterRegistryTypePass::class, 'CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\RegisterRegistryTypePass');
